third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-release-noref-expected.txt

EVENT(close)
EVENT(close)
This is a testharness.js-based test.
PASS MediaKeySession lifetime after release() without references 
Harness: the test ran to completion.
